+++ src/machine-state.h
@@ -67,7 +67,7 @@ struct machine_state {
     uint64_t fcsr;                       ///< CSR fcsr.
     std::array<uint64_t, F_REG_COUNT> f; ///< Floating-point register file.
 
-    uint64_t minstret; ///< CSR minstret.
+    uint64_t icycleinstret; ///< CSR icycleinstret.
 
     uint64_t mstatus;  ///< CSR mstatus.
     uint64_t mtvec;    ///< CSR mtvec.